Bugfix in Simulation.run
If the simulator tried to run for an amount of time that was smaller than the next simulated interaction, it would still simulate at least one interaction. This could lead to Simulation.simulator.t and Simulation.time becoming out of sync if the history_interval was sufficiently small.

The time_to_steps function used in Simulation.run has now been changed to fix this.